feat: add custom crud role and permission and multiple school admin

+++ app/Http/Controllers/Admin/AdminController.php
@@ -4,26 +4,31 @@ namespace App\Http\Controllers\Admin;
 
 use App\Models\Admin;
 use App\Models\School;
+use App\Models\AdminSchool;
 use Yajra\DataTables\DataTables;
 use Illuminate\Support\Facades\DB;
 use Spatie\Permission\Models\Role;
 use Illuminate\Support\Facades\Log;
 use App\Http\Controllers\Controller;
+use Illuminate\Support\Facades\Auth;
 use App\Http\Requests\Admin\AdminRequest;
 
 class AdminController extends Controller
 {
     public function __construct()
     {
-        $this->middleware(['permission:admin']);
+        // $this->middleware(['permission:admin']);
     }
     /**
      * Display a listing of the resource.
      */
     public function index()
     {
-        $data = Admin::with('roles')->latest()->get();
+        if (!Auth::user()->can('Manage Admin')) {
+            return redirect()->back()->with('error', 'Maaf, Anda tidak memiliki akses untuk halaman tersebut');
+        }
         if (request()->ajax()) {
+            $data = Admin::with('roles', 'adminSchool')->latest();
             return DataTables::of($data)
                 ->addColumn('role', function ($query) {
                     $role = "";
@@ -32,15 +37,22 @@ class AdminController extends Controller
                     }
                     return $role;
                 })
+                ->addColumn('school', function ($query) {
+                    $school = "";
+                    foreach ($query->adminSchool as $value) {
+                        $school .= "<span class='badge bg-success m-1'>{$value?->school?->name}</span>";
+                    }
+                    return $school;
+                })
                 ->addColumn('action', function ($data) {
                     $actionEdit = route('admin.edit', $data->id);
                     $actionDelete = route('admin.destroy', $data->id);
                     return "<div class='d-flex gap-2 flex-nowrap justify-content-center'>" .
-                        view('components.action.edit', ['action' => $actionEdit]) .
-                        view('components.action.delete', ['action' => $actionDelete, 'id' => $data->id]) .
+                        view('components.action.edit', ['action' => $actionEdit, 'name' => 'Admin']) . '&nbsp;' .
+                        view('components.action.delete', ['action' => $actionDelete, 'id' => $data->id, 'name' => 'Admin']) .
                         "</div>";
                 })
-                ->rawColumns(['action', 'role'])
+                ->rawColumns(['action', 'role', 'school'])
                 ->make(true);
         }
         return view('admins.admin.index');
@@ -51,9 +63,13 @@ class AdminController extends Controller
      */
     public function create()
     {
+        if (!Auth::user()->can('Create Admin')) {
+            return redirect()->back()->with('error', 'Maaf, Anda tidak memiliki akses untuk halaman tersebut');
+        }
         $roles = Role::get();
         $schools = School::orderBy('name')->get();
-        return view('admins.admin.create-edit', compact('roles', 'schools'));
+        $adminSchools = [];
+        return view('admins.admin.create-edit', compact('roles', 'schools', 'adminSchools'));
     }
 
     /**
@@ -61,6 +77,9 @@ class AdminController extends Controller
      */
     public function store(AdminRequest $request)
     {
+        if (!Auth::user()->can('Create Admin')) {
+            return redirect()->back()->with('error', 'Maaf, Anda tidak memiliki akses untuk halaman tersebut');
+        }
 
         try {
             DB::beginTransaction();
@@ -74,12 +93,21 @@ class AdminController extends Controller
 
             $admin = Admin::create($data);
             $admin->assignRole(Role::find($request->role_id)->name);
+            if ($request->admin_schools && is_array($request->admin_schools)) {
+                foreach ($request->admin_schools as $school) {
+                    AdminSchool::create([
+                        'admin_id' => $admin->id,
+                        'school_id' => $school
+                    ]);
+                }
+            }
 
             DB::commit();
 
             return redirect()->route('admin.index')->with('success', 'Berhasil menambah admin');
         } catch (\Exception $e) {
             DB::rollback();
+            dd($e->getMessage());
             Log::error($e->getMessage());
             // Handle the exception, log it, or return an error response
             return redirect()->back()->with('error', 'Gagal menambah admin: ' . $e->getMessage());
@@ -99,9 +127,13 @@ class AdminController extends Controller
      */
     public function edit(Admin $admin)
     {
+        if (!Auth::user()->can('Edit Admin')) {
+            return redirect()->back()->with('error', 'Maaf, Anda tidak memiliki akses untuk halaman tersebut');
+        }
         $roles = Role::get();
         $schools = School::orderBy('name')->get();
-        return view('admins.admin.create-edit', compact('admin', 'roles', 'schools'));
+        $adminSchools = $admin->adminSchool->pluck('school_id')->toArray();
+        return view('admins.admin.create-edit', compact('admin', 'roles', 'schools', 'adminSchools'));
     }
 
     /**
@@ -109,6 +141,9 @@ class AdminController extends Controller
      */
     public function update(AdminRequest $request, Admin $admin)
     {
+        if (!Auth::user()->can('Edit Admin')) {
+            return redirect()->back()->with('error', 'Maaf, Anda tidak memiliki akses untuk halaman tersebut');
+        }
         $data = $request->except('password');
 
         if (!empty($request->password)) {
@@ -124,6 +159,15 @@ class AdminController extends Controller
 
         $admin->update($data);
         $admin->syncRoles(Role::find($request->role_id)->name);
+        if ($request->admin_schools) {
+            $admin->adminSchool()->delete();
+            foreach ($request->admin_schools as $school) {
+                AdminSchool::create([
+                    'admin_id' => $admin->id,
+                    'school_id' => $school
+                ]);
+            }
+        }
         return redirect()->route('admin.index')->with('success', 'Berhasil mengubah data');
     }
 
@@ -132,7 +176,11 @@ class AdminController extends Controller
      */
     public function destroy(Admin $admin)
     {
+        if (!Auth::user()->can('Delete Admin')) {
+            return redirect()->back()->with('error', 'Maaf, Anda tidak memiliki akses untuk halaman tersebut');
+        }
         file_exists($admin->avatar) ? unlink($admin->avatar) : '';
+        $admin->adminSchool()->delete();
         $admin->delete();
         return redirect()->route('admin.index')->with('success', 'Berhasil menghapus admin');
     }
