+++ app/Http/Controllers/Admin/SchoolController.php
@@ -6,6 +6,7 @@ use App\Models\School;
 use Illuminate\Http\Request;
 use Yajra\DataTables\DataTables;
 use App\Http\Controllers\Controller;
+use Illuminate\Support\Facades\Auth;
 use App\Http\Requests\Admin\SchoolRequest;
 
 class SchoolController extends Controller
@@ -13,13 +14,16 @@ class SchoolController extends Controller
 
     public function __construct()
     {
-        $this->middleware('permission:school');
+        // $this->middleware('permission:school');
     }
     /**
      * Display a listing of the resource.
      */
     public function index()
     {
+        if (!Auth::user()->can('Manage Sekolah')) {
+            return redirect()->back()->with('error', 'Maaf, Anda tidak memiliki akses untuk halaman tersebut');
+        }
         if (request()->ajax()) {
             $data = School::latest()->get();
             return DataTables::of($data)
@@ -29,8 +33,8 @@ class SchoolController extends Controller
                     $actionShow = route('school.show', $data->id);
                     return "<div class='d-flex justify-content-center'>" .
                         view('components.action.show', ['action' => $actionShow, 'label' => 'Kelas']) .
-                        view('components.action.edit', ['action' => $actionEdit]) .
-                        view('components.action.delete', ['action' => $actionDelete, 'id' => $data->id]) .
+                        view('components.action.edit', ['action' => $actionEdit, 'name' => 'Sekolah']) .
+                        view('components.action.delete', ['action' => $actionDelete, 'id' => $data->id, 'name' => 'Sekolah']) .
                         "</div>";
                 })
                 ->rawColumns(['action'])
@@ -44,6 +48,9 @@ class SchoolController extends Controller
      */
     public function create()
     {
+        if (!Auth::user()->can('Create Sekolah')) {
+            return redirect()->back()->with('error', 'Maaf, Anda tidak memiliki akses untuk halaman tersebut');
+        }
         return view('admins.school.create-edit');
     }
 
@@ -52,6 +59,9 @@ class SchoolController extends Controller
      */
     public function store(SchoolRequest $request)
     {
+        if (!Auth::user()->can('Create Sekolah')) {
+            return redirect()->back()->with('error', 'Maaf, Anda tidak memiliki akses untuk halaman tersebut');
+        }
         School::create($request->validated());
         return redirect()->route('school.index')->with('success', 'Sekolah berhasil ditambahkan');
     }
@@ -61,16 +71,19 @@ class SchoolController extends Controller
      */
     public function show(string $id)
     {
+        if (!Auth::user()->can('Manage Sekolah')) {
+            return redirect()->back()->with('error', 'Maaf, Anda tidak memiliki akses untuk halaman tersebut');
+        }
         $school = School::findOrFail($id);
         if (request()->ajax()) {
-            $data = $school->classroom()->latest()->get();
+            $data = $school->classroom()->latest();
             return DataTables::of($data)
                 ->addColumn('action', function ($data) {
                     $actionEdit = route('classroom.edit', $data->id);
                     $actionDelete = route('classroom.destroy', $data->id);
                     return "<div class='d-flex justify-content-center'>" .
-                        view('components.action.edit', ['action' => $actionEdit]) .
-                        view('components.action.delete', ['action' => $actionDelete, 'id' => $data->id]) .
+                        view('components.action.edit', ['action' => $actionEdit, 'name' => 'Sekolah']) .
+                        view('components.action.delete', ['action' => $actionDelete, 'id' => $data->id, 'name' => 'Sekolah']) .
                         "</div>";
                 })
                 ->rawColumns(['action'])
@@ -84,6 +97,9 @@ class SchoolController extends Controller
      */
     public function edit(School $school)
     {
+        if (!Auth::user()->can('Edit Sekolah')) {
+            return redirect()->back()->with('error', 'Maaf, Anda tidak memiliki akses untuk halaman tersebut');
+        }
         return view('admins.school.create-edit', compact('school'));
     }
 
@@ -92,6 +108,9 @@ class SchoolController extends Controller
      */
     public function update(SchoolRequest $request, School $school)
     {
+        if (!Auth::user()->can('Edit Sekolah')) {
+            return redirect()->back()->with('error', 'Maaf, Anda tidak memiliki akses untuk halaman tersebut');
+        }
         $school->update($request->validated());
         return redirect()->route('school.index')->with('success', 'Sekolah berhasil diubah');
     }
@@ -101,6 +120,9 @@ class SchoolController extends Controller
      */
     public function destroy(School $school)
     {
+        if (!Auth::user()->can('Delete Sekolah')) {
+            return redirect()->back()->with('error', 'Maaf, Anda tidak memiliki akses untuk halaman tersebut');
+        }
         $school->delete();
         return redirect()->route('school.index')->with('success', 'Sekolah berhasil dihapus');
     }
